server/AlfredServer/requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.8.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication and security
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# Firebase
firebase-admin>=6.2.0

# AI and OpenAI
openai>=1.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2 